seprated addInput form to different component
