+++ src/components/index.ts
@@ -1,2 +1,3 @@
 export { Wrapper } from "./Wrapper";
 export { List } from "./List";
+export { AddInput } from "./AddInput";
